feature: moving WindowManagerToolsImplementation to a public-facing miral folder
